third_party/WebKit/LayoutTests/fast/forms/form-associated-element-expected.txt

Verify that only elements that are 'form associatable' get a form owner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS elements.length > 0 is true
Testing LABEL
PASS formOwner is "defined"
Testing INPUT
PASS formOwner is "defined"
Testing BUTTON
PASS formOwner is "defined"
Testing FIELDSET
PASS formOwner is "defined"
Testing INPUT
PASS formOwner is "defined"
Testing INPUT
PASS formOwner is "defined"
Testing KEYGEN
PASS formOwner is "defined"
Testing LABEL
PASS formOwner is "defined"
Testing INPUT
PASS formOwner is "defined"
Testing OBJECT
PASS formOwner is "defined"
Testing SELECT
PASS formOwner is "defined"
Testing TEXTAREA
PASS formOwner is "defined"
Testing IMG
PASS formOwner is "defined"
Testing OPTION
PASS formOwner is "not defined"
Testing DIV
PASS formOwner is "not defined"
Testing A
PASS formOwner is "not defined"
Testing P
PASS formOwner is "not defined"
Testing INPUT
PASS formOwner is "defined"
Testing KEYGEN
PASS formOwner is "defined"
Testing OBJECT
PASS formOwner is "defined"
Testing SELECT
PASS formOwner is "defined"
Testing PRE
PASS formOwner is "not defined"
Testing SPAN
PASS formOwner is "not defined"
Testing IMG
PASS formOwner is "not defined"
PASS successfullyParsed is true

TEST COMPLETE
